add vessels table to vessel group
