+++ apps/fishing-map/features/vessel-groups/VesselGroupAddButton.module.css
@@ -1,0 +1,18 @@
+.groupOptions {
+  padding: var(--space-S);
+  padding-right: 0;
+}
+
+.groupOption {
+  cursor: pointer;
+  padding: var(--space-XS);
+  padding-right: var(--space-M);
+}
+
+.groupOptionNew {
+  text-decoration: underline;
+}
+
+.groupOption:not(:last-of-type) {
+  border-bottom: var(--border);
+}
